Drawer fix and data test id addition to grid table
