+++ app/cdap/components/GridTable/components/GridTextCell/index.tsx
@@ -19,11 +19,11 @@ import React from 'react';
 import { useGridTextCellStyles } from './styles';
 import { IGridTextCellProps } from './types';
 
-export default function GridTextCell({ cellValue }: IGridTextCellProps) {
+export default function GridTextCell({ cellValue, cellIndex }: IGridTextCellProps) {
   const classes = useGridTextCellStyles();
 
   return (
-    <TableCell className={classes.tableRowCell}>
+    <TableCell className={classes.tableRowCell} data-testid={`grid-cellData-${cellIndex}`}>
       <Card className={classes.root} variant="outlined">
         <Typography className={classes.cell} data-testid={`grid-text-cell-${cellValue}`}>
           {cellValue}
