+++ app/cdap/components/GridTable/components/GridTextCell/types.ts
@@ -15,4 +15,5 @@
  */
 export interface IGridTextCellProps {
   cellValue: string;
+  cellIndex: string | number;
 }
