+++ app/cdap/components/WranglerGrid/AddTransformationPanel/SelectColumnsWidget/types.ts
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-import { IHeaderNamesList } from 'components/WranglerGrid/AddTransformationPanel/types';
+import { IHeaderNamesList } from 'components/WranglerGrid/SelectColumnPanel/types';
 
 export interface ISelectColumnsWidgetProps {
   selectedColumns: IHeaderNamesList[];
