+++ app/cdap/components/WranglerGrid/SelectColumnPanel/DataTable/CheckboxInput/__test__/CheckBoxInput.test.tsx
@@ -1,0 +1,53 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import { fireEvent, render, screen } from '@testing-library/react';
+import React from 'react';
+import CheckBoxInput from 'components/WranglerGrid/SelectColumnPanel/DataTable/CheckboxInput';
+
+describe('It should render ', () => {
+  it('Should render the checkBox component and click the checkbox with disable check false', () => {
+    const container = render(
+      <CheckBoxInput
+        selectedColumns={[{ label: 'test', type: [''], name: '' }]}
+        handleDisableCheckbox={() => false}
+        columnDetail={{ label: 'test', type: [''], name: '' }}
+        onMultipleSelection={() => jest.fn()}
+        columnIndex={0}
+      />
+    );
+
+    const checkboxInputElement = screen.getByTestId(/check-box-input-0/i);
+    fireEvent.click(checkboxInputElement);
+    expect(container).toBeDefined();
+  });
+
+  it('Should render the checkBox component and click the checkbox with disable checkbox as true', () => {
+    const container = render(
+      <CheckBoxInput
+        selectedColumns={[{ label: 'test', type: [''], name: '' }]}
+        handleDisableCheckbox={() => true}
+        columnDetail={{ label: 'test', type: [''], name: '' }}
+        onMultipleSelection={() => jest.fn()}
+        columnIndex={0}
+      />
+    );
+
+    const checkboxInputElement = screen.getByTestId(/check-box-input-0/i);
+    fireEvent.click(checkboxInputElement);
+    expect(container).toBeDefined();
+  });
+});
